* config/i386/i386.c (ix86_expand_sse_unpack): New.         * config/i386/i386-protos.h: Update.         * config/i386/sse.md (vec_unpacku_hi_v16qi, vec_unpacks_hi_v16qi,         vec_unpacku_lo_v16qi, vec_unpacks_lo_v16qi, vec_unpacku_hi_v8hi,         vec_unpacks_hi_v8hi, vec_unpacku_lo_v8hi, vec_unpacks_lo_v8hi,         vec_unpacku_hi_v4si, vec_unpacks_hi_v4si, vec_unpacku_lo_v4si,         vec_unpacks_lo_v4si): New.
        * gcc.dg/vect/vect-117.c (main1): Fix follow-on int/size_t
        mismatches.
        * gcc.dg/vect/vect-reduc-dot-u16.c: Update comments.

git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/autovect-branch@107679 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/vect/vect-117.c
@@ -16,13 +16,13 @@ int main1 ()
   /* Unknown evolution.  */
   for (i = 0; i < N; i++)
     {
-       ia[i] = (int *) *p[i];
+       ia[i] = (size_t *) *p[i];
     }
 
   /* check results: */
   for (i = 0; i < N; i++)
     {
-      if (ia[i] != (int *) *p[i])
+      if (ia[i] != (size_t *) *p[i])
          abort();
     }
   return 0;
